requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
html5lib==1.1
idna==3.4
lxml==4.9.1
python-slugify==6.1.2
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
text-unidecode==1.3
urllib3==1.26.12
webencodings==0.5.1
